requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
aiogram==2.11.2
    # via -r requirements.in
aiohttp==3.7.4.post0
    # via
    #   -r requirements.in
    #   aiogram
argcomplete==1.10.3
    # via textract
arrow==0.14.7
    # via ics
async-timeout==3.0.1
    # via aiohttp
attrs==21.4.0
    # via aiohttp
babel==2.10.1
    # via aiogram
beautifulsoup4==4.8.2
    # via textract
certifi==2021.10.8
    # via aiogram
chardet==3.0.4
    # via
    #   aiohttp
    #   pdfminer-six
    #   textract
compressed-rtf==1.0.6
    # via extract-msg
docx2txt==0.8
    # via textract
ebcdic==1.1.1
    # via extract-msg
extract-msg==0.28.7
    # via textract
ics==0.7
    # via -r requirements.in
idna==3.3
    # via yarl
imapclient==2.1.0
    # via extract-msg
lxml==4.8.0
    # via python-pptx
multidict==6.0.2
    # via
    #   aiohttp
    #   yarl
olefile==0.46
    # via extract-msg
pdfminer-six==20191110
    # via textract
pillow==9.1.0
    # via python-pptx
pycryptodome==3.14.1
    # via pdfminer-six
pydantic==1.9.0
    # via -r requirements.in
python-dateutil==2.8.2
    # via
    #   -r requirements.in
    #   arrow
    #   ics
python-pptx==0.6.21
    # via textract
pytz==2022.1
    # via babel
pytz-deprecation-shim==0.1.0.post0
    # via tzlocal
six==1.12.0
    # via
    #   ics
    #   imapclient
    #   pdfminer-six
    #   python-dateutil
    #   textract
sortedcontainers==2.4.0
    # via pdfminer-six
soupsieve==2.3.2.post1
    # via beautifulsoup4
speechrecognition==3.8.1
    # via textract
tatsu==5.6.1
    # via ics
textract==1.6.5
    # via -r requirements.in
typing-extensions==4.2.0
    # via
    #   aiohttp
    #   pydantic
tzdata==2022.1
    # via pytz-deprecation-shim
tzlocal==4.2
    # via extract-msg
xlrd==1.2.0
    # via textract
xlsxwriter==3.0.3
    # via python-pptx
yarl==1.7.2
    # via aiohttp
